# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Price Monitor API",
        "description": "API for monitoring product prices across different e-commerce platforms",
        "contact": {
            "name": "Price Monitor Support",
            "email": "support@pricemonitor.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get user alerts",
                "description": "Get all alerts for the authenticated user",
                "operationId": "getAlerts",
                "responses": {
                    "200": {
                        "description": "User alerts list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Create price alert",
                "description": "Create a new price alert for a product",
                "operationId": "createAlert",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alert created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/alerts/{alert}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get alert details",
                "description": "Get details of a specific alert",
                "operationId": "getAlert",
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "description": "Alert ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not the owner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Update alert",
                "description": "Update an existing price alert",
                "operationId": "updateAlert",
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "description": "Alert ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alert updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not the owner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Delete alert",
                "description": "Delete a price alert",
                "operationId": "deleteAlert",
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "description": "Alert ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Alert deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not the owner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/alerts/{alert}/should-trigger": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Check if alert should trigger",
                "description": "Check if the alert conditions are met",
                "operationId": "checkAlertTrigger",
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "description": "Alert ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert trigger check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertTriggerCheckResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not the owner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/alerts/{alert}/trigger": {
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Manually trigger alert",
                "description": "Manually trigger an alert",
                "operationId": "triggerAlert",
                "parameters": [
                    {
                        "name": "alert",
                        "in": "path",
                        "description": "Alert ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert triggered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertTriggerResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not the owner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/alerts/statistics": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Get alert statistics",
                "description": "Get system-wide alert statistics",
                "operationId": "getAlertStatistics",
                "responses": {
                    "200": {
                        "description": "Alert statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertStatistics"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "CreateAlertRequest": {
                "required": [
                    "product_id",
                    "condition",
                    "notification_channel"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "below",
                            "above",
                            "equals",
                            "percent_drop",
                            "percent_increase"
                        ],
                        "example": "below"
                    },
                    "target_price": {
                        "description": "Required for price-based conditions (below, above, equals)",
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "percent_threshold": {
                        "description": "Required for percentage-based conditions (percent_drop, percent_increase)",
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "notification_channel": {
                        "type": "string",
                        "enum": [
                            "email",
                            "webhook",
                            "sms",
                            "slack",
                            "discord",
                            "telegram"
                        ],
                        "example": "email"
                    }
                },
                "type": "object"
            },
            "UpdateAlertRequest": {
                "properties": {
                    "target_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1100
                    },
                    "percent_threshold": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "notification_channel": {
                        "type": "string",
                        "enum": [
                            "email",
                            "webhook",
                            "sms",
                            "slack",
                            "discord",
                            "telegram"
                        ],
                        "example": "email"
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "below",
                            "above",
                            "equals",
                            "percent_drop",
                            "percent_increase"
                        ],
                        "example": "below"
                    }
                },
                "type": "object"
            },
...
            "AlertResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/AlertWithProduct"
                    }
                },
                "type": "object"
            },
            "AlertCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertWithProduct"
                        }
                    }
                },
                "type": "object"
            },
...
            "AlertTriggerCheckResponse": {
                "properties": {
                    "should_trigger": {
                        "type": "boolean",
                        "example": true
                    },
                    "current_price": {
                        "type": "number",
                        "format": "float",
                        "example": 899.99
                    },
                    "target_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1000,
                        "nullable": true
                    },
                    "previous_price": {
                        "type": "number",
                        "format": "float",
                        "example": 1100,
                        "nullable": true
                    },
                    "condition": {
                        "type": "string",
                        "example": "below"
                    },
                    "message": {
                        "type": "string",
                        "example": "Alert conditions met"
                    }
                },
                "type": "object"
            },
            "AlertTriggerResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Alert triggered successfully"
                    },
                    "data": {
                        "properties": {
                            "trigger_count": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}